+++ src/validator.rs
@@ -4,14 +4,15 @@
 use anyhow::{Context, Result};
 use clap::{Args, ValueEnum};
 use clap_sys::version::clap_version_is_compatible;
-use regex::RegexBuilder;
+use rayon::prelude::*;
+use regex::{Regex, RegexBuilder};
 use serde::Serialize;
 use std::collections::BTreeMap;
 use std::fs;
 use std::path::PathBuf;
 use strum::IntoEnumIterator;
 
-use crate::plugin::library::PluginLibrary;
+use crate::plugin::library::{PluginLibrary, PluginMetadata};
 use crate::tests::{PluginLibraryTestCase, PluginTestCase, TestCase, TestResult, TestStatus};
 use crate::util;
 
@@ -19,7 +20,7 @@ use crate::util;
 /// [`tally()`][Self::tally()] method to compute the number of successful and failed tests.
 ///
 /// Uses `BTreeMap`s purely so the order is stable.
-#[derive(Debug, Serialize)]
+#[derive(Debug, Default, Serialize)]
 #[serde(rename_all = "kebab-case")]
 pub struct ValidationResult {
     /// A map indexed by plugin library paths containing the results of running the per-plugin
@@ -82,6 +83,12 @@ pub struct ValidatorSettings {
     /// Using this option will remove those protections, but in turn the tests may run faster.
     #[arg(long)]
     pub in_process: bool,
+    /// Run all tests in parallel.
+    ///
+    /// Some plugins may not handle this correctly, even when the plugins are loaded into different
+    /// processes. This cannot be used in combination with `--in-process`.
+    #[arg(short = 'j', long, conflicts_with = "in_process")]
+    pub parallel: bool,
 }
 
 /// Options for running a single test. This is used for the out-of-process testing method. This
@@ -117,11 +124,6 @@ pub enum SingleTestType {
 /// Run the validator using the specified settings. Returns an error if any of the plugin paths
 /// could not loaded, or if the plugin ID filter did not match any plugins.
 pub fn validate(settings: &ValidatorSettings) -> Result<ValidationResult> {
-    let mut results = ValidationResult {
-        plugin_library_tests: BTreeMap::new(),
-        plugin_tests: BTreeMap::new(),
-    };
-
     // Before doing anything, we need to make sure any temporary artifact files from the previous
     // run are cleaned up. These are used for things like state dumps when one of the state tests
     // fail. This is allowed to fail since the directory may not exist and even if it does and we
@@ -138,102 +140,209 @@ pub fn validate(settings: &ValidatorSettings) -> Result<ValidationResult> {
         })
         .transpose()?;
 
+    // The tests can optionally be run in parallel. This is not the default since some plugins may
+    // not handle it correctly, event when the plugins are loaded in different processes. It's also
+    // incompatible with the in-process mode.
     // TODO: We now gather all the results and print everything in one go at the end. This is the
     //       only way to do JSON, but for the human readable version printing things as we go could
     //       be nice.
-    for library_path in &settings.paths {
-        // We distinguish between two separate classes of tests: tests for an entire plugin library,
-        // and tests for a single plugin contained witin that library. The former group of tests are
-        // run first and they only receive the path to the plugin library as their argument, while
-        // the second class of tests receive an already loaded plugin library and a plugin ID as
-        // their arugmetns. We'll start with the tests for entire plugin libraries so the in-process
-        // mode makes a bit more sense. Otherwise we would be measuring plugin scanning time on
-        // libraries that may still be loaded in the process.
-        let mut plugin_library_results = Vec::new();
-        for test in PluginLibraryTestCase::iter() {
-            let test_name = test.to_string();
-            match (&test_filter_re, settings.invert_filter) {
-                (Some(test_filter_re), false) if !test_filter_re.is_match(&test_name) => continue,
-                (Some(test_filter_re), true) if test_filter_re.is_match(&test_name) => continue,
-                _ => (),
-            }
-
-            plugin_library_results.push(if settings.in_process {
-                test.run_in_process(library_path)
-            } else {
-                test.run_out_of_process(library_path, settings.hide_output)?
-            });
-        }
-
-        results
-            .plugin_library_tests
-            .insert(PathBuf::from(&library_path), plugin_library_results);
-
-        // And these are the per-plugin instance tests
-        let plugin_library = PluginLibrary::load(library_path)
-            .with_context(|| format!("Could not load '{}'", library_path.display()))?;
-        let metadata = plugin_library.metadata().with_context(|| {
-            format!(
-                "Could not fetch plugin metadata for '{}'",
-                library_path.display()
-            )
-        })?;
-        if !clap_version_is_compatible(metadata.clap_version()) {
-            log::debug!(
-                "'{}' uses an unsupported CLAP version ({}.{}.{}), skipping...",
-                library_path.display(),
-                metadata.version.0,
-                metadata.version.1,
-                metadata.version.2
-            );
-
-            continue;
-        }
-
-        // We only now know how many tests will be run for this plugin library
-        for plugin_metadata in metadata.plugins {
-            if results.plugin_tests.contains_key(&plugin_metadata.id) {
-                anyhow::bail!(
-                    "Duplicate plugin ID in validation results: '{}' ({}) has already been \
-                     validated",
-                    plugin_metadata.id,
-                    library_path.display(),
+    // TODO: There doesn't seem to be a way to run rayon iterators on the main thread, so the
+    //       parallel and scalar versions need to be duplicated here. We could also create a single
+    //       threaded shim that implements Rayon's parallel iterator methods, and then branch on the
+    //       places where we create parallel iterators instead.
+    let results = if settings.parallel {
+        settings
+            .paths
+            .par_iter()
+            .map(|library_path| {
+                // We distinguish between two separate classes of tests: tests for an entire plugin
+                // library, and tests for a single plugin contained witin that library. The former
+                // group of tests are run first and they only receive the path to the plugin library
+                // as their argument, while the second class of tests receive an already loaded
+                // plugin library and a plugin ID as their arugmetns. We'll start with the tests for
+                // entire plugin libraries so the in-process mode makes a bit more sense. Otherwise
+                // we would be measuring plugin scanning time on libraries that may still be loaded
+                // in the process.
+                let mut plugin_library_tests: BTreeMap<PathBuf, Vec<TestResult>> = BTreeMap::new();
+                plugin_library_tests.insert(
+                    library_path.clone(),
+                    PluginLibraryTestCase::iter()
+                        .par_bridge()
+                        .filter(|test| test_filter(test, settings, &test_filter_re))
+                        .map(|test| {
+                            if settings.in_process {
+                                Ok(test.run_in_process(library_path))
+                            } else {
+                                test.run_out_of_process(library_path, settings.hide_output)
+                            }
+                        })
+                        .collect::<Result<Vec<TestResult>>>()?,
                 );
-            }
 
-            // It's possible to filter by plugin ID in case you want to validate a single plugin
-            // from a plugin library containing multiple plugins
-            match &settings.plugin_id {
-                Some(plugin_id) if &plugin_metadata.id != plugin_id => continue,
-                _ => (),
-            }
+                // And these are the per-plugin instance tests
+                let plugin_library = PluginLibrary::load(library_path)
+                    .with_context(|| format!("Could not load '{}'", library_path.display()))?;
+                let plugin_metadata = plugin_library.metadata().with_context(|| {
+                    format!(
+                        "Could not fetch plugin metadata for '{}'",
+                        library_path.display()
+                    )
+                })?;
+                if !clap_version_is_compatible(plugin_metadata.clap_version()) {
+                    log::debug!(
+                        "'{}' uses an unsupported CLAP version ({}.{}.{}), skipping...",
+                        library_path.display(),
+                        plugin_metadata.version.0,
+                        plugin_metadata.version.1,
+                        plugin_metadata.version.2
+                    );
 
-            let mut plugin_test_results = Vec::new();
-            for test in PluginTestCase::iter() {
-                let test_name = test.to_string();
-                match (&test_filter_re, settings.invert_filter) {
-                    (Some(test_filter_re), false) if !test_filter_re.is_match(&test_name) => {
-                        continue
-                    }
-                    (Some(test_filter_re), true) if test_filter_re.is_match(&test_name) => continue,
-                    _ => (),
+                    // Since this is a map-reduce, this acts like a continue statement in a loop. We
+                    // could use `.filter_map()` instead but that would only make things more
+                    // complicated
+                    return Ok(ValidationResult::default());
                 }
 
-                plugin_test_results.push(if settings.in_process {
-                    test.run_in_process((&plugin_library, &plugin_metadata.id))
-                } else {
-                    test.run_out_of_process(
-                        (&plugin_library, &plugin_metadata.id),
-                        settings.hide_output,
-                    )?
-                });
-            }
+                // We only now know how many tests will be run for this plugin library. We'll count
+                // the number of plugins that match the filters and then compare that against the
+                // number of entries in the map to make sure there are no dupli
+                let plugin_tests: BTreeMap<String, Vec<TestResult>> = plugin_metadata
+                    .plugins
+                    .into_par_iter()
+                    .filter(|plugin_metadata| plugin_filter(plugin_metadata, settings))
+                    // We're building a `BTreeMap` containing the results for all plugins in the
+                    // plugin's library
+                    .map(|plugin_metadata| {
+                        Ok((
+                            plugin_metadata.id.clone(),
+                            PluginTestCase::iter()
+                                .par_bridge()
+                                .filter(|test| test_filter(test, settings, &test_filter_re))
+                                .map(|test| {
+                                    if settings.in_process {
+                                        Ok(test
+                                            .run_in_process((&plugin_library, &plugin_metadata.id)))
+                                    } else {
+                                        test.run_out_of_process(
+                                            (&plugin_library, &plugin_metadata.id),
+                                            settings.hide_output,
+                                        )
+                                    }
+                                })
+                                .collect::<Result<Vec<TestResult>>>()?,
+                        ))
+                    })
+                    .collect::<Result<BTreeMap<_, _>>>()?;
 
-            results
-                .plugin_tests
-                .insert(plugin_metadata.id, plugin_test_results);
-        }
-    }
+                Ok(ValidationResult {
+                    plugin_library_tests,
+                    plugin_tests,
+                })
+            })
+            .reduce(
+                || Ok(ValidationResult::default()),
+                |a, b| {
+                    // Monads galore! The fact that we need to handle errors for plugin tests makes
+                    // this a bit more complicated.
+                    let (a, b) = (a?, b?);
+
+                    // In the serial version this could be done when iterating over the plugins, but
+                    // when using iterators you can't do that. But it's still essential to make sure
+                    // we don't test two versionsq of the same plugin.
+                    if a.intersects(&b) {
+                        anyhow::bail!(
+                            "Duplicate plugin ID in validation results. Maybe multiple versions \
+                             of the same plugin are being validated."
+                        );
+                    }
+
+                    Ok(ValidationResult::union(a, b))
+                },
+            )
+    } else {
+        settings
+            .paths
+            .iter()
+            .map(|library_path| {
+                let mut plugin_library_tests: BTreeMap<PathBuf, Vec<TestResult>> = BTreeMap::new();
+                plugin_library_tests.insert(
+                    library_path.clone(),
+                    PluginLibraryTestCase::iter()
+                        .filter(|test| test_filter(test, settings, &test_filter_re))
+                        .map(|test| {
+                            if settings.in_process {
+                                Ok(test.run_in_process(library_path))
+                            } else {
+                                test.run_out_of_process(library_path, settings.hide_output)
+                            }
+                        })
+                        .collect::<Result<Vec<TestResult>>>()?,
+                );
+
+                let plugin_library = PluginLibrary::load(library_path)
+                    .with_context(|| format!("Could not load '{}'", library_path.display()))?;
+                let plugin_metadata = plugin_library.metadata().with_context(|| {
+                    format!(
+                        "Could not fetch plugin metadata for '{}'",
+                        library_path.display()
+                    )
+                })?;
+                if !clap_version_is_compatible(plugin_metadata.clap_version()) {
+                    log::debug!(
+                        "'{}' uses an unsupported CLAP version ({}.{}.{}), skipping...",
+                        library_path.display(),
+                        plugin_metadata.version.0,
+                        plugin_metadata.version.1,
+                        plugin_metadata.version.2
+                    );
+
+                    return Ok(ValidationResult::default());
+                }
+
+                let plugin_tests: BTreeMap<String, Vec<TestResult>> = plugin_metadata
+                    .plugins
+                    .into_iter()
+                    .filter(|plugin_metadata| plugin_filter(plugin_metadata, settings))
+                    .map(|plugin_metadata| {
+                        Ok((
+                            plugin_metadata.id.clone(),
+                            PluginTestCase::iter()
+                                .filter(|test| test_filter(test, settings, &test_filter_re))
+                                .map(|test| {
+                                    if settings.in_process {
+                                        Ok(test
+                                            .run_in_process((&plugin_library, &plugin_metadata.id)))
+                                    } else {
+                                        test.run_out_of_process(
+                                            (&plugin_library, &plugin_metadata.id),
+                                            settings.hide_output,
+                                        )
+                                    }
+                                })
+                                .collect::<Result<Vec<TestResult>>>()?,
+                        ))
+                    })
+                    .collect::<Result<BTreeMap<_, _>>>()?;
+
+                Ok(ValidationResult {
+                    plugin_library_tests,
+                    plugin_tests,
+                })
+            })
+            .reduce(|a, b| {
+                let (a, b) = (a?, b?);
+
+                if a.intersects(&b) {
+                    anyhow::bail!(
+                        "Duplicate plugin ID in validation results. Maybe multiple versions of \
+                         the same plugin are being validated."
+                    );
+                }
+
+                Ok(ValidationResult::union(a, b))
+            })
+            .unwrap_or_else(|| Ok(ValidationResult::default()))
+    }?;
 
     if let Some(plugin_id) = &settings.plugin_id {
         if results.plugin_tests.is_empty() {
@@ -280,6 +389,32 @@ pub fn run_single_test(settings: &SingleTestSettings) -> Result<()> {
     })
 }
 
+/// The filter function for determining whether or not a test should be run based on the validator's
+/// settings settings.
+fn test_filter<'a, T: TestCase<'a>>(
+    test: &T,
+    settings: &ValidatorSettings,
+    test_filter_re: &Option<Regex>,
+) -> bool {
+    let test_name = test.to_string();
+    match (&test_filter_re, settings.invert_filter) {
+        (Some(test_filter_re), false) if !test_filter_re.is_match(&test_name) => false,
+        (Some(test_filter_re), true) if test_filter_re.is_match(&test_name) => false,
+        _ => true,
+    }
+}
+
+/// The filter function for determining whether or not tests should be run for a particular plugin.
+fn plugin_filter(plugin_metadata: &PluginMetadata, settings: &ValidatorSettings) -> bool {
+    // It's possible to filter by plugin ID in case you want to validate a single plugin
+    // from a plugin library containing multiple plugins
+    #[allow(clippy::match_like_matches_macro)]
+    match &settings.plugin_id {
+        Some(plugin_id) if &plugin_metadata.id != plugin_id => false,
+        _ => true,
+    }
+}
+
 impl ValidationResult {
     /// Count the number of passing, failing, and skipped tests.
     pub fn tally(&self) -> ValidationTally {
@@ -307,6 +442,33 @@ impl ValidationResult {
             num_skipped,
             num_warnings,
         }
+    }
+
+    // Check whether the maps in the object intersect. Useful to ensure that a plugin ID only occurs
+    // once in the outputs before merging them.
+    pub fn intersects(&self, other: &Self) -> bool {
+        for key in other.plugin_library_tests.keys() {
+            if self.plugin_library_tests.contains_key(key) {
+                return true;
+            }
+        }
+
+        for key in other.plugin_tests.keys() {
+            if self.plugin_tests.contains_key(key) {
+                return true;
+            }
+        }
+
+        false
+    }
+
+    /// Merge the results from two validation result objects. If `other` contains a key that also
+    /// exists in this object, then the version from `other` is used.
+    pub fn union(mut self, other: Self) -> Self {
+        self.plugin_library_tests.extend(other.plugin_library_tests);
+        self.plugin_tests.extend(other.plugin_tests);
+
+        self
     }
 }